Dosya düzenleme
-Gereksiz scriptler kaldırıldı
