+++ src/socket/client-with-stream.py
@@ -1,0 +1,20 @@
+#! /usr/bin/python3
+
+import socket
+
+s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+host ="127.0.0.1"
+port =8000
+s.connect((host,port))
+
+def ts(str):
+   s.send('client gönderdi'.encode()) 
+   data = ''
+   data = s.recv(1024).decode()
+   print (data)
+
+while 2:
+   r = input('enter')
+   ts(s)
+
+s.close ()
